ms-auth/authentification/pyproject.toml
[project]
name = "authentification"
version = "0.1.0"
description = ""
authors = [
    {name = "Lee-Rudy",email = "120602756+Lee-Rudy@users.noreply.github.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.30",
    "psycopg[binary]>=3.2.0",
    "PyJWT>=2.9.0",
    "bcrypt>=4.1.3",
]

[tool.poetry]
packages = [{include = "authentification", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-cov = ">=5.0.0"
httpx = ">=0.27.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
